backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt==4.0.1",
    "fastapi[all]>=0.116.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pip>=25.1.1",
    "psycopg[binary,c]>=3.2.9",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.35.0",
]
